deflection implemented and folder restructure
reorganized some of my folders. deflection implemented - created new health script for the player called
HealthScript (same as the old one but with the new features implemented). deal damage script added which was used to test this - attach to object with collider to do damage at regular intervals while overlapping.